requirements.txt
asn1crypto==0.24.0
azure-common==1.1.14
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.4
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-storage-common==1.3.0
azure-storage-nspkg==3.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
class-registry==2.1.2
cryptography==2.3.1
filters==1.3.2
futures==3.1.1
idna==2.7
pip==18.1
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
pysha3==1.0.2
python-dateutil==2.7.3
pytz==2018.5
regex==2018.8.17
requests==2.19.1
setuptools==39.0.1
six==1.11.0
urllib3==1.23
